AutoDuckBase in lib-builtin-effects
