+++ libraries/lib-builtin-effects/CMakeLists.txt
@@ -5,6 +5,8 @@ Library of built-in effects. Implement the UI-agnostic APIs of lib-effects.
 set( SOURCES
    AmplifyBase.cpp
    AmplifyBase.h
+   AutoDuckBase.cpp
+   AutoDuckBase.h
    BassTrebleBase.cpp
    BassTrebleBase.h
    ChangePitchBase.cpp
